8.0 regression - major overhaul

+++ clean.ps1
@@ -1,7 +1,10 @@
 # Clean out all files from output paths
 . ".\harness-data.ps1"
 if ("Y" -eq (Read-Host "Enter Y to clear out all output files")) {
-   foreach ($d in ($DATA.outputPaths.GetEnumerator())) {
+   foreach ($d in ($DATA.filePaths.GetEnumerator())) {
+      if (@("licenses","tests","diagnostics") -contains $d.Key) {
+         continue
+      }
       if ((Test-Path $d.Value -PathType Container) -and "Y" -eq (Read-Host "Enter Y to nuke everything in $($d.Value)")) {
          Write-Host "Clearing .\$(Split-Path -Path $($d.Value) -Leaf)"
          Remove-Item -Recurse -Path "$($d.Value)\*"
